Removed NameBasedHashCodeHolder from joints and rigid-body.

+++ ihmc-humanoid-robotics/src/main/java/us/ihmc/humanoidRobotics/communication/wholeBodyTrajectoryToolboxAPI/WholeBodyTrajectoryToolboxAPI.java
@@ -7,5 +7,5 @@ import us.ihmc.sensorProcessing.frames.ReferenceFrameHashCodeResolver;
 
 public interface WholeBodyTrajectoryToolboxAPI<M>
 {
-   void set(M message, Map<Long, RigidBody> rigidBodyNamedBasedHashMap, ReferenceFrameHashCodeResolver referenceFrameResolver);
+   void set(M message, Map<Integer, RigidBody> rigidBodyNamedBasedHashMap, ReferenceFrameHashCodeResolver referenceFrameResolver);
 }
